a little Progress on GUI
I Added some funktions on Buttons and Played around with the GUI. Now it
kind of looks like the way I imagned it in the 1st place.
It added an Icon from iconfinder.com
I added the "FontAwsome" to the Buildtree because I intend to use some
of its Symbols later. (http://fortawesome.github.io/Font-Awesome/)